refactor: extracted the payment handler logic from the contorller
